requirements.txt
asgiref==3.9.2
cloudpickle==3.1.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.9.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
pillow==11.3.0
PyJWT==2.10.1
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
